+++ cmake/share/Files.cmake
@@ -12,8 +12,8 @@ endif()
 
 include( PreFiles OPTIONAL)
 
-include( "cmake/Headers.cmake" OPTIONAL)
-include( "cmake/Sources.cmake" OPTIONAL)
+include( Headers OPTIONAL)
+include( Sources OPTIONAL)
 
 include_directories( ${INCLUDE_DIRS})
 
@@ -23,12 +23,15 @@ set( PROJECT_FILES
    ${PROJECT_FILES}
    ${SOURCES}
    ${PUBLIC_HEADERS}
+   ${PUBLIC_GENERATED_HEADERS}
    ${PRIVATE_HEADERS}
+   ${PRIVATE_GENERATED_HEADERS}
    ${CMAKE_EDITABLE_FILES}
 )
 
 set( PROJECT_INSTALLABLE_HEADERS
    ${PUBLIC_HEADERS}
+   ${PUBLIC_GENERATED_HEADERS}
    ${PRIVATE_HEADERS}
 )