+++ cmake/share/Headers.cmake
@@ -1,11 +1,19 @@
+### If you want to edit this, copy it from cmake/share to cmake. It will be
+### picked up in preference over the one in cmake/share. And it will not get
+### clobbered with the next upgrade.
+
+if( MULLE_TRACE_INCLUDE)
+   message( STATUS "# Include \"${CMAKE_CURRENT_LIST_FILE}\"" )
+endif()
+
 #
 # The following includes include definitions generated
-# during `mulle-sde update`. Don't edit those files. They are
+# during `mulle-sde reflect`. Don't edit those files. They are
 # overwritten frequently.
 #
 # === MULLE-SDE START ===
 
-include( _Headers)
+include( _Headers OPTIONAL)
 
 # === MULLE-SDE END ===
 #
@@ -16,29 +24,20 @@ include( _Headers)
 # MULLE_MATCH_TO_CMAKE_HEADERS_FILE="DISABLE" # or NONE
 #
 
-#
-# Add ignored headers back in so that the generators pick them up
-#
-set( PUBLIC_HEADERS
-"src/_mulle-objc-runtime-include.h"
-${PUBLIC_HEADERS}
-)
 
 # keep headers to install separate to make last minute changes
-set( INSTALL_PUBLIC_HEADERS ${PUBLIC_HEADERS})
+set( INSTALL_PUBLIC_HEADERS ${PUBLIC_HEADERS}
+${PUBLIC_GENERATED_HEADERS}
+)
 
 #
 # Do not install generated private headers and include-private.h
 # which aren't valid outside of the project scope.
 #
 set( INSTALL_PRIVATE_HEADERS ${PRIVATE_HEADERS})
-list( REMOVE_ITEM INSTALL_PRIVATE_HEADERS "include-private.h")
-
-# add ignored headers back in so that the generators pick them up
-set( PRIVATE_HEADERS
-"src/_mulle-objc-runtime-include-private.h"
-${PRIVATE_HEADERS}
-)
+if( INSTALL_PRIVATE_HEADERS)
+   list( REMOVE_ITEM INSTALL_PRIVATE_HEADERS "include-private.h")
+endif()
 
 #
 # You can put more source and resource file definitions here.