+++ test/.mulle/share/env/environment-plugin-os-darwin.sh
@@ -1,5 +1,8 @@
 #
 # Git mirror and Zip/TGZ cache to conserve bandwidth
 #
-export MULLE_FETCH_MIRROR_DIR="${HOME:-/tmp}/Library/Caches/mulle-fetch/git-mirror"
-export MULLE_FETCH_ARCHIVE_DIR="${HOME:-/tmp}/Library/Caches/mulle-fetch/archive"
+MULLE_FETCH_MIRROR_DIR="${MULLE_FETCH_MIRROR_DIR:-${HOME:-/tmp}/Library/Caches/mulle-fetch/git-mirror}"
+export MULLE_FETCH_MIRROR_DIR
+
+MULLE_FETCH_ARCHIVE_DIR="${MULLE_FETCH_ARCHIVE_DIR:-${HOME:-/tmp}/Library/Caches/mulle-fetch/archive}"
+export MULLE_FETCH_ARCHIVE_DIR