+++ RELEASENOTES.md
@@ -1,3 +1,7 @@
+### 0.17.1
+
+* new mulle-sde project structure
+
 ## 0.17.0
 
 * experimental stackdepth indent for method trace (not that satisfactory, but occasionally useful)